+++ midilib/src/main/cpp/oscillators/SawtoothOscillator.cpp
@@ -1,6 +1,6 @@
 #include "SawtoothOscillator.h"
 
-float SawtoothOscillator::calculate(float time, float frequency, int8_t overtoneFactor) {
+float SawtoothOscillator::calculate(double time, float frequency, int8_t overtoneFactor) {
     float p = calcPhase(time, frequency, overtoneFactor);
     return mAmplitude * float(p / M_PI);
 }
