+++ midilib/src/main/cpp/oscillators/SinOscillator.cpp
@@ -4,6 +4,6 @@
 #include "oboe/Oboe.h"
 #endif
 
-float SinOscillator::calculate(float time, float frequency, int8_t overtoneFactor) {
+float SinOscillator::calculate(double time, float frequency, int8_t overtoneFactor) {
     return mAmplitude * sinf(calcPhase(time, frequency, overtoneFactor));
 }
