+++ midilib/src/main/cpp/oscillators/TriangleOscillator.h
@@ -8,7 +8,7 @@
 class TriangleOscillator: public Oscillator {
     using Oscillator::Oscillator;
 
-    float calculate(float time, float frequency, int8_t overtoneFactor) override;
+    float calculate(double time, float frequency, int8_t overtoneFactor) override;
 };
